+++ src/apps/DDBEncounterMunch.js
@@ -3,7 +3,7 @@ import utils from "../lib/utils.js";
 import CompendiumHelper from "../lib/CompendiumHelper.js";
 import PatreonHelper from "../lib/PatreonHelper.js";
 import MuncherSettings from "../lib/MuncherSettings.js";
-import { getAvailableCampaigns } from "../lib/Settings.js";
+import { getAvailableCampaigns } from "../lib/DDBCampaigns.js";
 import AdventureMunchHelpers from "../muncher/adventure/AdventureMunchHelpers.js";
 import { importCharacterById } from "./DDBCharacterManager.js";
 import SETTINGS from "../settings.js";