+++ src/parser/DDBEncounters.js
@@ -2,7 +2,7 @@
 import logger from "../logger.js";
 import FileHelper from "../lib/FileHelper.js";
 import { getCobalt } from "../lib/Secrets.js";
-import { getAvailableCampaigns } from "../lib/Settings.js";
+import { getAvailableCampaigns } from "../lib/DDBCampaigns.js";
 import SETTINGS from "../settings.js";
 import DDBProxy from "../lib/DDBProxy.js";
 import DDBEncounterMunch from "../apps/DDBEncounterMunch.js";