+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # NEXT UP
 
 - Default door sound added to DDB Maps.
+- Some code refactoring
+- Allow users to sync character back to DDB using their own patreon key, this can be set in the Advanced tab of the character.
 
 # 3.4.53